refact: pageSort파일 내부 함수 분리 및 네이밍 변경
