+++ src/components/page-layout-ui/PersonalPageContentSection.tsx
@@ -12,7 +12,7 @@ import { usePageStore, useParentsFolderIdStore } from '@/stores/pageStore';
 import { FolderDetail } from '@/types/folders';
 import { LinkDetail } from '@/types/links';
 import { AddLinkModalSkeleton } from '../skeleton/AddLinkModalSkeleton';
-import { sortPageData } from '@/utils/pageDataManage';
+import { sortPageData } from '@/utils/handlePageDataSort';
 import { usePageDragAndDrop } from '@/hooks/usePageDragAndDrop';
 import { useDragAndDropSensors } from '@/utils/dragAndDrop';
 import MobileFolderCard from '../folder-card/mobile/MobileFolderCard';
