+++ src/components/page-layout-ui/SharedPageFolderContentSection.tsx
@@ -10,7 +10,7 @@ import { usePageStore, useParentsFolderIdStore } from '@/stores/pageStore';
 import { LinkDetail } from '@/types/links';
 import { FolderDetail } from '@/types/folders';
 import { AddLinkModalSkeleton } from '../skeleton/AddLinkModalSkeleton';
-import { sortPageData } from '@/utils/pageDataManage';
+import { sortPageData } from '@/utils/handlePageDataSort';
 import { usePageDragAndDrop } from '@/hooks/usePageDragAndDrop';
 import { useDragAndDropSensors } from '@/utils/dragAndDrop';
 import MobileFolderCard from '../folder-card/mobile/MobileFolderCard';
