+++ src/pages/BookmarkPage.tsx
@@ -4,7 +4,7 @@ import PageHeaderSection from '@/components/page-layout-ui/PageHeaderSection';
 import PageControllerSection from '@/components/page-layout-ui/PageControllerSection';
 import useFetchFavorite from '@/hooks/queries/useFetchFavorite';
 import { usePageSort } from '@/hooks/usePageSort';
-import { getPageDataLength } from '@/utils/pageDataManage';
+import { getPageDataLength } from '@/utils/getPageDataLength';
 import { PageLayout } from '@/components/common-ui/PageLayout';
 import { useMobile } from '@/hooks/useMobile';
 import { baseCards } from '@/constants/homeCards';
