+++ src/pages/SharedPage.tsx
@@ -4,7 +4,7 @@ import { useFetchSharedPage } from '@/hooks/queries/useFetchSharedPage';
 import { usePageStore, useParentsFolderIdStore } from '@/stores/pageStore';
 import { usePageSort } from '@/hooks/usePageSort';
 import { useMobile } from '@/hooks/useMobile';
-import { getPageDataLength } from '@/utils/pageDataManage';
+import { getPageDataLength } from '@/utils/getPageDataLength';
 import { PageLayout } from '@/components/common-ui/PageLayout';
 import ScrollToTopButton from '@/components/common-ui/ScrollToTopButton';
 import PageHeaderSection from '@/components/page-layout-ui/PageHeaderSection';
